Switch to mingw-64. Support 64-bit mingw builds. Support VS2013. Remove nmake support. (first part)
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1884 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/Common/Win32/gdi_print.c
@@ -872,7 +872,7 @@ gdi_screenful(struct ea *ea, unsigned short rows, unsigned short cols,
 		    ((pstate.usable_rows - 1) * pstate.space_size.cy) -
 		    pchar.poffY,
 		0, NULL,
-		pstate.caption, strlen(pstate.caption), NULL);
+		pstate.caption, (UINT)strlen(pstate.caption), NULL);
 	if (status <= 0) {
 	    *fail = "ExtTextOut failed";
 	    rc = -1;
@@ -1255,13 +1255,13 @@ get_printer_device(const char *printer_name, HGLOBAL *pdevnames,
     assert(dn);
     memset(dn, '\0', sizeof(DEVNAMES));
     offset = sizeof(DEVNAMES);
-    dn->wDriverOffset = offset;
+    dn->wDriverOffset = (WORD)offset;
     memcpy((char *)dn + offset, pi->pDriverName, ldn);
     offset += ldn;
-    dn->wDeviceOffset = offset;
+    dn->wDeviceOffset = (WORD)offset;
     memcpy((char *)dn + offset, pi->pPrinterName, lpn);
     offset += lpn;
-    dn->wOutputOffset = offset;
+    dn->wOutputOffset = (WORD)offset;
     memcpy((char *)dn + offset, pi->pPortName, ltn);
     dn->wDefault = 0;
